datasets/nclt_kaggle/requirements.txt
# Core
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Point Cloud Processing
open3d>=0.17.0
MinkowskiEngine>=0.5.4

# Data & Config
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Evaluation
scipy>=1.10.0
scikit-learn>=1.2.0

# SLAM (optional, install separately)
# evo  # trajectory evaluation

# Development
pytest>=7.3.0
black>=23.3.0
mypy>=1.3.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
